use local type guard for isObservable
NPM's `is-observable` depends on a pre-ES-module version of `symbol-observable`.  When used with a modern version, it checks for `value[{ default: $$observable }]` (which obviously fails).  By implementing this type guard locally, we can correct these issues.

Moreover, `is-observable` is itself a CommonJS module.  As Igor Minor described in his ReactiveConf 2017 talk, CommonJS dependencies deoptimize Webpack, yielding bigger bundles.  This tradeoff isn't worth it for a 1-line function that can be trivially reimplemented.