Add empty comments to make eslint happy with prettier
Prettier was formatting the css onto the same line because it is short,
however eslint was then showing errors for the 'no-useless-concat' rule.

The code appears to have been formatted onto multiple lines to make it
look somewhat close to CSS, so forced the formatting to multiple line
with an empty comment.

An alternative to concatenating multple string for the tests would be to
write a tagged template literal function that removes all the
indentation so code still looks nice, and the tests pass.

Example:

    expect(sheets.toString()).to.be(stripIndent`
        .my-test-class {
          color: red;
        }`);