Check prettier status in pre-commit hook
Want to enforce like eslint.

Removed `git add` because it wasn't doing anything as it was. If the
files were being modified with `eslint --fix` or `prettier --write` then
the auto changed files would be added and the commit updated. However
this doesn't work with partially staged files so it ends up causing more
pain than it's worth.

If users install the prettier plugin for their favourite editor and
enable Format on Save then they should have no issues at all, and the
`git add` part of the lint-staged hook is not needed.