Simplify flow distribution
Since we already have all types in `src` we can reuse them instead of
copying. And we don't even need a tool for that.

I add src to published files and generate this file with simple `echo`
```js
// @flow

export * from '../src';
```

Prevented starting server on flow check which could lead to performance
problems and crashing in CI.