Migrate to yarn
One of NPM maintainers shows some disrespect to others which kills a
trust to them.

Yarn is more stable client with more full set of features so I suggest
to migrate to it.

As you can see yarn.lock is two times smaller because of less verbose
syntax. Locfile is not required for every package in monorepo, thanks to
yarn workspaces.