Line 23 grammar mistake (#1633)
Old:
1. Does JSS has overhead?
New:
1. Does JSS have overhead?
Why?
When the subject of a sentence is a third person (he, she, it) we need to follow the rules of Subject-Verb Agreement. Basically add an s to the end of the verb, but when the verb "do" is used it has priority over the other verb so it always matches the subject. In this case the verb "have" has also been matched by mistake.